Ensure XML is valid before importing
This commit adds a validator class that uses a RelaxNG schema
to ensure that the XML is valid before importing. The XML
must have a `<metadata>` element with a <collection_name> and then
one or more `<record>` elements. The record elements require a
`<title>` element.

This commit also fixes an issue with the method that removes the
author string from titles. Before the title would need to have a
slash in it because the string was being split on the slash, but
now the method will check to see if the title has a slash in it
before splitting. This was causing some records to fail when
importing the Sargent and Sims collection.