Story #127: Download a zip of a work's files
This commit is a work in progress.  Changes in this commit:

Added a WorkZip class that knows how to build the zip file and keeps
track of which background job will be called to build the zip.  It will
also keep track of the location of the zip file once it has been built.

Copied the code to build the zip from the BulkDownload class, which I
intend to delete later.

Handled the edge case of adding files with duplicate names into the zip.