Added link to download a zip or button to create a zip. Story #127
Work in progress.  Changes in this commit:

Copied the work show view from hyrax so that we could add in the UI for
the zip file.

Added UI to download a zip file if it's available, or build a new zip
file if it isn't available.  If the background job is still working, it
will display a message for the user to check back in a few minutes.

Added the active_job_status and redis-activesupport gems to make it
easier to check the status of the background job.