Initial models for bagging
This commit contains a `Bag` and `BulkBag` class
that are used for creating individual bags
and for creating multiple bags from an array of ids.

The bags that are generated by `Bag` are tar
files. This is because they will eventually need
to be downloaded/transferred. tar files are less
complex than zip files and don't compress files.

We don't need compression because the Work contents
is going to be binary data (PDFs and images). DPN
and APTrust both require tarred bags.

https://wiki.aptrust.org/Bagging_specifications
https://wiki.duraspace.org/display/DPNC/BagIt+Specification++OLD

These will be used in a rake task in later work so
that they can be used from the command line.

Connected to #191