Add dotenv-rails
This will allow the application to load environment variables
from a file in the Rails root directory called .env.ENVIRONMENT.
For example, in production, environment variables can be added
to a file called .env.production. This is the way DCE does server
builds these days, but the FRBM servers were built before this
practice was put in place. Adding dotenv shouldn't change any
behavior for the existing application, except that it will allow
it to run on newly built DCE servers.