UI for creating bags
This adds the ability to create bags
from the 'My Works' section of the dashboard.

Select the works you want to bag, and then click
on the 'Add to Bag' button.

You will be redirected to notifications with a notice
that you will need to refresh that page for updates on
the completion of the bagging.

When the bagging is done, you'll get a notification.

This commit adds the `xray-rails` gem as a way to determine
what views need to be edited.

It also changes the capybara selenium driver to use chrome and
updates the `travis.yml` so that it will install the chrome addon.

Connected to #188