Added check for current page to feature spec.
This feature spec was reliably passing for me locally, but failed on one
of the CI builds.

I was able to reproduce the failure locally like this:
rspec spec/features --seed 42993

I ran a debugger, and watched the controller properly queue the job, but
the spec still failed saying that the job was not queued.

I'm able to get the spec passing by adding another check ahead of the
check for the queued job.  I don't know why this works. I don't think
this is a problem with the code, but rather a problem with the timing of
the spec (timing for rspec or capybara or the test job queue).