Do not save searches from crawler bots
We believe that crawler bots are filling up the database, and that then
when the delete saved searches cron job runs that is causing out of
memory problems because there are so many saved searches to delete.
There's no reason to be storing crawler bot saved searches in the first
place, so let's see if this fixes the problem.