Extract method to determine publication year
The date slider expects an integer year, but our source
data might have a variety of strings with date-like data in it.

The new method extracts the first sequence of four digits it finds
and treats it as the publication year.

The test show how the method functions for a variety of possible
input data.