Simplify image derivative processing (#564)
**ISSUES**
The image validation spec is dependent upon imagemagick being installed
in the test environment which increases setup complexity and time in
the CI environment.

The ImageValidator class duplicates funcitonality built-into
ImageMagick without offering any additonal functionality.

The override of Hyrax::CharacterizeJob calls an undefined class named
Cypripedium::CorruptFileError.

This bug isn't caught because there's no test coverage of the local
version of Hyrax::CharacterizeJob

**RESOLUTION**
Refactor the CharacterizeJob to call MiniMagick using a method
that fails gracefully when ImageMagic is not installed.  Then, update
the error code to log a simple string describing the error instead of
calling a non-existent class.

We still lack test coverage for CharacterizeJob, but the overall solution
is simpler and no longer calls non-existent classes.