Allow first item on multi-value form fields to be removed
**ISSUE**
For non-editable form fields like 'creator', there was no way to
change the first entry once it was saved.

**FIX**
Update the remove buttons to show on all but the last item in
multi-item lists.  I.E. if there is only one item, it is also the
last item, so don't allow it to be deleted.