Update Capybara for compatibility with update Puma version
**ISSUE**
The update to Puma causes a handful of test in the test suite to fail
due to depencency inconsistencies in the test suite.

**FIX**
Update Capybara and related dependencies to allow test suite to run
successfully.

NOTE: This change and the previous change have no production impact.
The production environment uses Apache rather than Puma as the
web server.