Show creator names in the JSON show view for works
**ISSUE**
The application stores data for `creator_id` pointing to the
corresponding items in the Creator authority.  The `creator` field,
which would usually contain the display names, is left blank.

**FIX**
Explicitly call a lookup method that populates the `creator`
attribute with display names corresponding to the authority IDs
stored in the `creator_id` attribute.