Implement a modal citations dialog
**STORY**
As a user, I want to remain in the context of the work I'm viewing
while I access citations, instead of being taken to a different page
that does not have navigation links back to my work.

**SOLUTION**
Implement citations as a modal pop up on the work's main show page
instead of linking to a separate citation page.

NOTES:
1. Users could get back to the work page using their browser back
button, but the previous UI did not provide clear navigational
hints.
2. This fix implements Hyrax's default citations for Chicago, MLA,
and APA styles.  Additional customizations are required to update
the citation formatters to display Cypripedium specific metadata.

REFERENCES:
This work was highly influenced by:
* Citation work by Brad Watson at Emory Libraries https://github.com/emory-libraries/dlp-lux/commit/006688cd05e2c9361eb018f0424616d4903e1ce9
* Bootstrap documentation https://getbootstrap.com/docs/5.2/components/modal/
* This CodePen by Moshe Assayag https://codepen.io/mozez/pen/rYXBvZ