Fix database schema and migration chain
**ISSUE**
There was a migration included in the code that was recently reverted.
Additionally, the schema file had manual edits that didn't correspond
to any migrations.

We want to return the schema to a clean state and ensure that the
migration chain is reversible for servers that have had the reverted
migration deployed.

**SOLUTION**
Re-instantiate the unwanted migration and add an explicit migration
to undo its changes.  This allows us to apply the migrations to any
production-like environment without having to manually rollback.

This commit also includes a 'clean' schema that removes any manual
edits.