Refactor URL presenter to work when request is missing
**ISSUE**
In our tests, we had to create a fake request object just to provide the
host for the URL presenter.  We'd like to be able to configure tests
involving work presenters without requring a dummy request.

**SOLUTION**
Keep the existing code which handles production requests and allows
us to specify a specific host in tests when desired.

Add a fallback that returns only the relative path (only_path: true)
when a valid request object isn't present.