Index numerically sortable volume and issue number values
**ISSUE**
The `issue_number_ssi` field is indexed as a string, so it sorts
lexically (alphabetically) instead of numerically.  This leads to
surprising order like:
        "Vol. 4, No. 3"
        "Vol. 4, No. 4"
        "Vol. 40, No. 1"
        "Vol. 41, No. 1"
        "Vol. 42, No. 1"
        "Vol. 5, No. 1"
        "Vol. 5, No. 2"

**RESOLUTION**
Parse the `issue_number` string into independent integer values
for the volume (if present) and issue.  Index these numeric values
so that we can later update the sort to use them instead of the
text sort.

The values above can then be sorted as expected:
    volume: 4   issue: 3
    volume: 4   issue: 4
    volume: 5   issue: 1
    volume: 5   issue: 2
    volume: 40  issue: 1
    volume: 41  issue: 1
    volume: 42  issue: 1