Update issue sort to use numerical values
**ISSUE**
Using the `issue_number` field for sorting results in an alphabetical
sort where numbers are orderd alphabetically - e.g.
```
  "1"
  "10"
  "105"
  "2"
  "40"
  "5"
```

**FIX**
Now that we have numeric `volume_number_isi` and `issue_number_isi`
fields in the index, we can sort numerically - e.g.
```
   1
   2
   5
   10
   40
   105
```

This commit updates the sort options to use the new numeric fields.

It also adds the system create time to the sort to ensure that results
are displaed in a deterministic order if the volume and isssue numbers
are equal for two items (also when they are missing).