Display creator names in JSON for individual works
**ISSUE**
Creator names were not being displayed when rendering an individual
work in the repository. The creator_ids were displayed, but because
creator names are not persisted to fedora, they are not included in
the Hyrax::Work object that is retrieved from Fedora.

**RESOLUTION**
Replace the creator node with the presenter(i.e. solr_document) value
for alpha_creator.

**TODO**
For simplicity and performance, we may want to either persist the
creator names to Fedora and/or base the JSON view on the presenter
object instead of the Hyrax::Work