[MAINTENANCE] Refactor TarBag to remove global variables
**ISSUE**
A number of instance variables were used to pass values across
methods which made code linkages and dependencies difficult
to analyze.

**RESOLUTION**
This change removes all instance variables in the class and
passes required data as paramaters to the method calls.
Where data from the instance initialization is required,
we've replaced references to instance variables with
corresponding attribute readers.

This change also adds more detailed testing of the TAR file
that gets created to confirm it has the esxpected content.