[REFACTOR] Make deployment info in footer easier to test
**RATIONALE**
Logic for multiple environments in the initializer made it difficult
to test because the initializer runs before the test suite and
configures values with settings for the test environment.

**RESOLUTION**
Extract the environment detection and configuration logic to a
utility class so that we can stub the relevant environment settings
and test code pathways for all environments.