Simplify overrides for using FITS servlet
**ISSUE**
The application copied all of the Hydra::Works::CharacterizationService
class in order to conditionally override one line in the `#extract_metadata`
method:

```
  def extract_metadata(content)
    Hydra::FileCharacterization.characterize(content, file_name, tools) do |cfg|
      cfg[:fits] = Hydra::Derivatives.fits_path
    end
  end
```

Gets overriden if the FITS servled URL is configured
```
  def extract_metadata(content)
    if ENV['FITS_SERVLET_URL']
      Hydra::FileCharacterization.characterize(content, file_name, :fits_servlet)
    else
      Hydra::FileCharacterization.characterize(content, file_name, :fits) do |cfg|
        cfg[:fits] = Hydra::Derivatives.fits_path
      end
    end
  end
```

**RESOLUTION**
We can actually push the check for the servlet downstream and override less code by
modifying the `#command` method in the Hydra::FileCharacterization::Characterizers::Fits
characterizer class.

This allows us to specifically target and test only the code we want to modify,
without pulling extra methods or classes that we don't need (or want) to override.