[FIX] Display full deprecation messages
**ISSUE**
We are not getting details of the deprecation warnings.

The cbeer/deprecation gem uses an old calling signature
for warnings, but Hyrax and related code expect to use the calling
signature in ActiveSupport::Deprecation. So, the warn method
was reading the message as the context and setting the message to
nil, giving a very generic message.

**FIX**
Check to see whether the first argument being passed is a string,
which means the call is using the two-parameter calling signature,
or a class or other object, which means the call is using the
three-parameter calling signature.

**BEFORE**
>DEPRECATION WARNING: You are using deprecated behavior which will be removed from the next major or minor release. (called from block in <top (required)> at /Users/mark/_no_backup_/Projects/cypripedium/config/initializers/hyrax.rb:47) (1 times); e.g.:
>
>DEPRECATION WARNING: You are using deprecated behavior which will be removed from the next major or minor release. (called from new at /Users/mark/_no_backup_/Projects/cypripedium/app/lib/rdf_xml_service.rb:9) (2 times); e.g.:

**AFTER**
>DEPRECATION WARNING: google_analytics_id is deprecated; use analytics_id from config/analytics.yml instead. (called from block in <top (required)> at /Users/mark/_no_backup_/Projects/cypripedium/config/initializers/hyrax.rb:47) (1 times); e.g.:
>
>DEPRECATION WARNING: passing `request` is deprecated; pass a host as `hostname:` instead. (called from new at /Users/mark/_no_backup_/Projects/cypripedium/app/lib/rdf_xml_service.rb:9) (2 times); e.g.: