Default to Docker for Solr and Fedora
**ISSUE**
The solr_wrapper and fcrepo_wrapper gems are no longer maintained.
Docker containers are the preferred method to run Solr and
Fedora services in the local development and test environments.

**RESOLUTION**
This commit updates the Docker configuration to support running
Solr and Fedora as containerized services. We've also removed any
obsolete configuration related to solr_wrapper and fcrepo_wrapper.