Upgrade to Hyrax 5.x and Ruby 3.x
This commit contains all the changes necessary to upgrade
the Hyrax engine from v3.6.0 to v5.0.1.

The majority of changes are to dependency versions along with
the necessary config file changes for moving from
* Ruby 2.7.4 -> 3.2.4
* Rails 5.2.8.1 -> 6.1.7.10
* Blacklight 6.25.0 -> 7.40.0
* Hyrax 3.6.0 -> 5.0.1

With this commit, the test suite runs successfuly and
the web application will load and run. Due to the update
from Bootstrap 3.x to Bootsrap 4.x, there are a number of
visual styling changes and problems that still need to be
addressed in a separate change.