Fix new work creation button
**ISSUE**
At some point in the upgrade process, the new work creation
workflow ceased to work because the "Add New Work" button would
not trigger the appropriate link. The new work selection modal
uses javascript to determine which worktype to route
to when creating new works.

**DAIGNOSIS**
The issue appears to be load-order related, so we tested
various permutations in `application.js` until we found
a working order.

**RESOLUTION**
This change adds a test for the workflow and updates JS load
order to ensure the necessary scripts get attached to the
selection dialog box modal.