Fix broken new work creation workflow
**ISSUE**
After the upgrade to Hyrax v5.0.4 the "Add New Work" button in the
administrator dashboard stopped working. This prevents repository
administrators from creating new works via the dashboard UI.

**DIAGNOSIS**
The problem was tracked down to an issue with the Javascript load
order and timeout issues due to asynchronous Javascript initialization.

**RESOULTION**
The timeout issues were addressed in https://github.com/curationexperts/cypripedium/pull/658.

This change adds a test that emulates the failing UI workflow, and
modifies the Javascript load order as required to fix the problem.