Update Sidekiq to latest release
**ISSUE**
Since upgrading to Hyrax 5, we are experiencing issues with
background jobs after deploying to production. Thumbnails
(i.e. derivative files) are not being attached
to newly created works.

**RATIONALE**
Before attempting to debug the issue further, we want to ensure we
are running the most recent compatible version of Sidekiq to
check whether that resolves the issues we're seeing.

We've noticed that the sidekiq service is not restarting (i.e.
reloading with updated application code) when deploying to prod-like
environments using Capistrano. We have a hypothesis that we may need
to re-generate the Systemd service files for Sidekiq.