Fix broken homepage link to collections
**ISSUE**
The "View All Collections" button on the hompage no longer displays
a list of collections, it takes the user to a search result page
showing all items in the repository.

**RESOLUTION**
Configure an inaccessible facet for the model ("Type") and update the
link to a search for all items with model == Collection.