Remove expensive filter added to all Solr queries
**ISSUE**
Catalog searches in Hyrax 5 are taking over three times longer than
in Hyrax 3.

**DIAGNOSIS**
The default processor chain used by the CatalogController includes
a very expensive filter that limits which collections are returned
for the collection facet.

**RESOLUTION**
Cypripedium does not display a collection facet, so we can remove the
filter from the processing chain.