Add Ruby version to repo
Adding a version file for ruby managers allows us to mange the
upgrade process more easily when deploying to production and
prod-like servers.

If we have rbenv and the corresponding ruby versions installed
on there server, we can deploy code that requires a different
ruby version without having to build a new server or taking the
server out of service so we can update the ruby version.