Remove Passenger from local Gemfile
**ISSUE**
We're getting erratic server behavior that appears to be caused by
conflicts between the system Passenger (installed via Ansible) and
the local Passenger version (configured via Gemfile). It appears that
Capistrano is restarting a copy of the local version instead of
reloading the system version at the end of each deployment.

**RESOLUTION**
Remove Passenger from the Gemfile and rely on the version installed
system-wide on the server.