Update to Hyrax 5.1 and Rails 7
**RATIONALE**
We are experiencing file versioning issues that may be addressed
by fixes in Hyrax v5.1.0. Even if the release does not address
the specific issues we're seeing, we want our fixes to apply to the
latest release so that we don't have to re-work them at some
future point.

**NOTE**
This change also includes small changes for Rails 7 compatibiility.