Ensure credentials are shared across deployments
We want Capstrano to treat `./config/master.key` as a shared file
so it persists across deployments.

With this strategy, the `master.key` file must be installed out
of band from the deployment process, but should not need to be
updated unless we choose to re-encode the credentials with a
new key in the future.