Link show page fields to facets
**ISSUES**
Links to creators, series, subjects, and resource types were linking
to a search on the fields instead of the corresponding facet. This
lead to confusing faceting appearance and filter terms in the UI.

**Resolution**
This change links show page links to the corresponding facet whenever
possible. It also leverages internationalization to ease future
rollout of more complete translations.