Set copyright year when application starts
**ISSUE**
The copyright displayed in the application footer was out-of-date
and displayed different years depending on the language.

**DIAGNOSIS**
The copright year was being set in the individual translation files
and not being updated consistently.

**RESOLUTION**
This change sets a global constant once when the application starts
and uses that constant to populate the copyright year value for
all internationalization strings.