Fix handling of non-parseable creation dates
**ISSUE**
Some works were not displaying because the citation presenter code
was encountering unexpected values that could not successfully be
parsed. Unexpected date values resulted in an unhandled exceptions
instead of having graceful failure modes.

**RESOLUTION**
Add a test case reprensenting the data we saw in the live application
and refactor the code to handle unexpected values gracefully:
In this case, we use the deposit year if the creation date is missing
or can not be parsed.