Add group field to Creators to allow us to group reports
**ISSUE**
We want to be able to report on author (Creator) publication
statistics grouped by author status - e.g. staff, consultant, other.

**RESOLUTION**
Add a group field to the Creator class. This field defaults to
'unassigned', but can also have the values 'staff' or 'consultant'.

**NOTE**
This change also refactors the creator model tests to avoid
creating unnecessary database objects and to test validations
directly instead of checking for exceptions on object creation.