Fix flaky tests
**ISSUE**
This system test is failing in about 1 of every 5 test runs:
```
  3) New work creation administrative deposit
     Failure/Error: choose 'Publication'

     Capybara::ElementNotFound:
       Unable to find visible radio button "Publication" that is not disabled

     [Screenshot Image]: /Users/mark/_no_backup_/Projects/cypripedium/tmp/capybara/failures_r_spec_example_groups_new_work_creation_administrative_deposit_996.png

     # ./spec/system/new_work_workflow_spec.rb:23:in `block (2 levels) in <top (required)>'
```

**DIAGNOSIS**
The work-type-selection modal dialog is styled with a CSS fade-in
transition (e.g. start viible: false, end visible: true)
that appears not to be completing reliably to allow Capybara to
consistently find the expected elements.

**Resolution**
Disable all CSS animations and transitions in the test environment.

**References**
* https://www.reddit.com/r/rubyonrails/comments/15n53qx/disable_animations_to_stabilize_browser_tests/
* https://marcgg.com/blog/2015/01/05/css-animations-failing-capybara-specs/