Leave cells blank instead of 0 in author report
**ISSUE**
The update to use date_uploaded_dtsi as the basis for author stats
reporting added zero-counts to the results so that there would
not be any gaps in reporting periods. This had the side effect of
displaying 0 for periods where authors didn't publish any works,
instead of leaving them blank as previously.

The extra 0s make the table much harder to scan for acutal publications,

**REOLUTION**
We introduced code to suppress zeros in author counts and
leave the cell blank instead.