Restore catalog search results JSON view to Blacklight 6.x format
**ISSUE**
Blacklight 7+ updated JSON views to render more complex attribute
structures. This breaks current integrations with SiteCore, which
enable recently updated Research Database documents to be displayed
on the main website.

**RESLUTION**
Add a test of the JSON rendering to check sample structure and content
to ensure it conforms to the Sitecore interface expectations. Then
locally override the search results (catalog/index) builder to
generate results consistent with previous versions of Blacklight and
Cypripedium.